chore: New BaseWebViewActivity
