+++ src/main/AndroidManifest.xml
@@ -5,13 +5,13 @@
 
     <application>
         <activity
-            android:name=".web_views.CheckoutActivity"
+            android:name=".web_views.widgets.CheckoutActivity"
             android:exported="false" />
         <activity
-            android:name=".web_views.ElementsActivity"
+            android:name=".web_views.widgets.ElementsActivity"
             android:exported="false" />
         <activity
-            android:name=".web_views.PaymentWidgetActivity"
+            android:name=".web_views.widgets.PaymentWidgetActivity"
             android:exported="false" />
         <activity
             android:name=".web_views.base.SubWebViewActivity"
