+++ src/main/res/layout/webview_activity.xml
@@ -3,7 +3,7 @@
     android:id="@+id/deuna_webview_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".web_views.base.BaseWebViewActivity">
+    tools:context=".web_views.DeunaWebViewActivity">
 
     <WebView
         android:id="@+id/deuna_webview"
